% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart.R
\name{ladBarChart}
\alias{ladBarChart}
\title{LAD Bar Chart}
\usage{
ladBarChart(data, var, highlight = NULL)
}
\arguments{
\item{data}{The data.}

\item{var}{The statistic to plot.}

\item{highlight}{The LADs to highlight. A vector.}
}
\description{
Create an ordered bar chart for a given statistic and highlight selected
regions
}
\examples{
\dontrun{
ladBarChart(emp, "measure_a", c("Cornwall", "Newport"))
}

}
