% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapChoicesUI.R
\name{mapChoicesUI}
\alias{mapChoicesUI}
\title{Set up the map choice buttons.}
\usage{
mapChoicesUI(id, select)
}
\arguments{
\item{id}{The name for the shiny id.}

\item{select}{The default variable to be selected.}
}
\description{
Set up the map choice buttons.
}
