% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatPlot}
\alias{scatPlot}
\title{Comparison of Composite Variables}
\usage{
scatPlot(data, x, y, colour, key, highlight = NULL, xLab = NULL,
  yLab = NULL, discrete = TRUE)
}
\arguments{
\item{data}{The employment data, see details for more information.}

\item{x}{The variable to be plotted on the x-axis.}

\item{y}{The variable to be plotted on the y-axis.}

\item{colour}{The column to colour the points by.}

\item{key}{The key for the tooltip.}

\item{highlight}{A vector of Local Authority District names to highlight on
the plot.}

\item{xLab}{The x-axis label.}

\item{yLab}{The y-axis label.}

\item{discrete}{Logical. Whether the data are discrete or continuous.}
}
\description{
Create a scatter plot of the composite variables
}
\details{
The \code{data} needs to contain \code{measureA}, \code{measureB} and the
employment rate statistics.
}
\author{
Nathan Eastwood
}
